Lots of work on the algorithm scheduling and scoring the target orbs. Still needs tweaking, but it's getting there.

+++ Kobold2D-2.0.3/HoopyDrop/Projectfiles/CollisionHandler.h
@@ -15,7 +15,7 @@
 -(void) setBody: (b2Body*) body;
 -(b2Body*) body;
 -(void) handleCollision: (b2Body*) body;
--(double) createTime;
+-(int) createTime;
 -(bool) isRemoved;
 -(void) markRemoved;
 -(void) removeThisTarget;
